fastMRI tutorial

+++ tensorflow_mri/python/models/conv_blocks.py
@@ -174,7 +174,8 @@ class ConvBlock(tf.keras.Model):
                kernel_initializer=self._kernel_initializer,
                bias_initializer=self._bias_initializer,
                kernel_regularizer=self._kernel_regularizer,
-               bias_regularizer=self._bias_regularizer))
+               bias_regularizer=self._bias_regularizer,
+               dtype=self.dtype))
       if self._use_batch_norm:
         self._norms.append(
             bn(axis=self._channel_axis,
